refactor

+++ src/main/java/fu/swp/dorm_mnm/model/User.java
@@ -23,7 +23,7 @@ public class User implements UserDetails {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id;
 
   @Column(name = "username", nullable = false, unique = true, length = 20)
   private String username;
@@ -64,13 +64,12 @@ public class User implements UserDetails {
   @Column(name = "updated_at")
   private Date updatedAt;
 
-  // @Enumerated(EnumType.STRING)
   @ManyToOne
   @JoinColumn(name = "role_id")
   private Role role;
 
-//  @OneToMany(mappedBy = "user")
-//  private List<Token> tokens;
+  // @OneToMany(mappedBy = "user")
+  // private List<Token> tokens;
 
   @Override
   public Collection<? extends GrantedAuthority> getAuthorities() {
